fix(#4591): extract suite-detection into a shipped module
gsd-test caught a real packaging violation immediately after the
suite-filter fix landed: tests/packaging-shipped-scripts-require-only-shipped.test.cjs
failed -- scripts/gen-platform-conformance-tier.cjs (which ships, per real
`npm pack --dry-run --json` resolution) required ./run-tests.cjs (which
deliberately does NOT ship -- it's a repo-development-only tool), producing
MODULE_NOT_FOUND in a real published install.

Fixed by extracting MARKED_SUITES + suiteOf (pure relocation, zero logic
change) into a new scripts/lib/suite-detection.cjs module, which DOES ship
(everything under scripts/lib/ ships). run-tests.cjs now imports and
re-exports suiteOf from there unchanged -- every existing consumer
(scripts/affected-tests-lib.cjs, tests/ci-test-scope.test.cjs) keeps working
byte-identically. gen-platform-conformance-tier.cjs now imports the shared
module directly, never touching run-tests.cjs.

Proactively (not waiting for a third gsd-test failure) also regenerated the
19 golden install-tree fixtures and registered the new file in
bin/install.js's GSD_SCRIPTS_LIB_FILES -- the same two ripples already hit
twice for the first new shipped file in this PR.
