+++ tests/fixtures/install-tree/pi.json
@@ -426,5 +426,6 @@
   "scripts/lib/ndjson-reporter.cjs",
   "scripts/lib/npm-version-check-diagnosis.cjs",
   "scripts/lib/platform-conformance-tier.generated.cjs",
-  "scripts/lib/shellcheck-fetch.cjs"
+  "scripts/lib/shellcheck-fetch.cjs",
+  "scripts/lib/suite-detection.cjs"
 ]
